define int long longの黒魔術をやめる。サイトによっては色々と不整合が起こり面倒なので

+++ library/cpp/array/lis.test.cpp
@@ -3,7 +3,7 @@
 
 signed main(){
     int n;cin>>n;
-    vector<int> a(n);
+    vector<ll> a(n);
     rep(i, n) {
         cin>>a[i];
     }
